Create forked rn_fusebox entry point
