+++ config/gni/devtools_grd_files.gni
@@ -304,6 +304,7 @@ grd_files_release_sources = [
   "front_end/entrypoints/main/main.js",
   "front_end/entrypoints/ndb_app/ndb_app.js",
   "front_end/entrypoints/node_app/node_app.js",
+  "front_end/entrypoints/rn_fusebox/rn_fusebox.js",
   "front_end/entrypoints/rn_inspector/rn_inspector.js",
   "front_end/entrypoints/shell/shell.js",
   "front_end/entrypoints/wasmparser_worker/wasmparser_worker-entrypoint.js",
@@ -570,6 +571,7 @@ grd_files_release_sources = [
   "front_end/panels/web_audio/web_audio.js",
   "front_end/panels/webauthn/webauthn-meta.js",
   "front_end/panels/webauthn/webauthn.js",
+  "front_end/rn_fusebox.html",
   "front_end/rn_inspector.html",
   "front_end/services/puppeteer/puppeteer.js",
   "front_end/services/tracing/tracing.js",
