+++ front_end/BUILD.gn
@@ -26,6 +26,7 @@ group("front_end") {
     "entrypoints/lighthouse_worker:bundle",
     "entrypoints/ndb_app:entrypoint",
     "entrypoints/node_app:entrypoint",
+    "entrypoints/rn_fusebox:entrypoint",
     "entrypoints/rn_inspector:entrypoint",
     "entrypoints/shell",
     "entrypoints/wasmparser_worker:worker_entrypoint",
@@ -112,6 +113,7 @@ node_action("html_entrypoints") {
     "js_app",
     "ndb_app",
     "node_app",
+    "rn_fusebox",
     "rn_inspector",
     "worker_app",
     "device_mode_emulation_frame",
