+++ front_end/entrypoints/rn_fusebox/README.md
@@ -1,0 +1,7 @@
+# `rn_fusebox.html`
+
+`rn_fusebox.html` is a customized entry point designed to work well with React Native targets. This entry point is designed to work with all features of the next-gen debugging stack (codename: React Fusebox ⚡).
+
+## Notes on embedding
+
+* Adding `&sources.hide_add_folder=true` to the inspector frontend URL hides UI related to Filesystem / Workspaces that does not work in hosted mode.
