+++ front_end/entrypoints/shell/BUILD.gn
@@ -39,6 +39,7 @@ devtools_entrypoint("shell") {
     "../js_app:*",
     "../ndb_app:*",
     "../node_app:*",
+    "../rn_fusebox:*",
     "../rn_inspector:*",
     "../worker_app:*",
   ]
