+++ front_end/index.html
@@ -20,6 +20,9 @@
       <a href="rn_inspector.html">rn_inspector.html ⚛️</a>
     </li>
     <li>
+      <a href="rn_fusebox.html">rn_fusebox.html ⚛️</a>
+    </li>
+    <li>
       <a href="inspector.html">inspector.html</a>
     </li>
     <li>
