+++ front_end/panels/recorder/BUILD.gn
@@ -74,6 +74,7 @@ devtools_entrypoint("meta") {
     "../..:*",
     "../../entrypoints/devtools_app:*",
     "../../entrypoints/inspector:*",
+    "../../entrypoints/rn_fusebox:*",
     "../../entrypoints/rn_inspector:*",
   ]
 }
